requirements.txt
streamlit
langchain-community
sentence-transformers
langchain-groq
python-dotenv
